Use classes for JavaScript component wrappers
