+++ app/view_components/example_iterations_summary_table.rb
@@ -1,0 +1,8 @@
+class ExampleIterationsSummaryTable < ViewComponent
+  def render(solution)
+    react_component("example-iterations-summary-table", {
+                      solution_id: solution.id,
+                      iterations: solution.serialized_iterations
+                    })
+  end
+end
