+++ app/view_components/maintaining_iterations_summary_table.rb
@@ -1,0 +1,7 @@
+class MaintainingIterationsSummaryTable < ViewComponent
+  def render(iterations)
+    react_component("maintaining-iterations-summary-table", {
+                      iterations: iterations.map(&:serialized)
+                    })
+  end
+end
