+++ app/view_components/notification_icon.rb
@@ -1,0 +1,5 @@
+class NotificationIcon < ViewComponent
+  def render(user)
+    react_component("notification-icon", { count: user.notifications.count })
+  end
+end
