+++ app/views/test/components/example/iterations_summary_table/index.html.haml
@@ -1,1 +1,1 @@
-= example_iterations_summary_table(@solution)
+= ExampleIterationsSummaryTable.new.render(@solution)
