+++ app/views/test/components/maintaining/iterations_summary_table/index.html.haml
@@ -1,1 +1,1 @@
-= maintaining_iterations_summary_table(@iterations)
+= MaintainingIterationsSummaryTable.new.render(@iterations)
