+++ test/view_components/view_component_test_case.rb
@@ -1,0 +1,10 @@
+require "test_helper"
+
+class ViewComponentTestCase < ActionView::TestCase
+  def assert_component_equal(component, params)
+    assert_dom_equal(
+      %(<div data-react-#{params[:id]}="true" data-react-data=#{params[:props].to_json} />),
+      component
+    )
+  end
+end
